Added action ticking to enemies. Refactored shit to make the Enemy struct smaller. Enemies can shoot
Adding a Timer field was causing the game to crash!? So I slimmed it down a bit

+++ src/game_gui.odin
@@ -5,10 +5,10 @@ import rl "vendor:raylib"
 // Draw functions are called at the end of each frame by the game.
 // This function draws the player's health and xp bar
 draw_game_gui :: proc(using game : ^Game) {
-    health_bar_width    := min(player.max_hth * 2, f32(rl.GetScreenWidth() - 30))
-    health_bar_x        := i32((f32(rl.GetScreenWidth()) - health_bar_width) / 2)
+    health_bar_width    := min(i32(player.max_hth * 2), rl.GetScreenWidth() - 30)
+    health_bar_x        := i32((f32(rl.GetScreenWidth()) - f32(health_bar_width)) / 2)
     health_bar_y        := rl.GetScreenHeight() - 50
-    health_width        := health_bar_width * (player.hth / player.max_hth)
+    health_width        := i32(f32(health_bar_width) * (f32(player.hth) / f32(player.max_hth)))
     rl.DrawRectangle(health_bar_x, health_bar_y, i32(health_width), 8, rl.RED)
     rl.DrawRectangleLines(health_bar_x, health_bar_y, i32(health_bar_width), 8, rl.WHITE)
